# drip-manager/swagger.yaml
swagger: "2.0"
info:
  description: "The Dynamic Real-time infrastructure planner (DRIP) allows application developers to seamlessly plan a customized virtual infrastructure based on application level constraints on QoS and resource budgets, provisioning the virtual infrastructure, deploy application components onto the virtual infrastructure, and start execution on demand using TOSCA."
  version: "3.0.0"
  title: "DRIP"
  contact:
    email: "z.zhao@uva.nl"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/drip-api"
schemes:
- "https"
- "http"
paths:
  /tosca_template:
    post:
      summary: "upload a tosca template description file"
      description: ""
      operationId: "uploadToscaTemplate"
      consumes:
      - "multipart/form-data"
      parameters:
      - in: "formData"
        name: "file"
        type: file
        description: "tosca Template description"
        required: true
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        405:
          description: "Invalid input"
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
  /tosca_template/ids:
    get:
      summary: "Get all topolog template IDs"
      description: "Returns all IDss "
      operationId: "getToscaTemplateIDs"
      produces:
        - "application/json"
      responses:
        200:
          description: "successful operation"
          schema:
            type: array
            items:
              type: string
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"        
  /tosca_template/{id}:
    get:
      summary: "Find topolog template by ID"
      description: "Returns a single topolog template"
      operationId: "getToscaTemplateByID"
      produces:
      - "text/plain"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to return"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
    put:
      summary: "Updates exisintg topolog template"
      description: ""
      operationId: "updateToscaTemplateByID"
      consumes:
      - "multipart/form-data"
      produces:
      - "text/plain"
      parameters:
      - in: "formData"
        name: "file"
        type: file
        description: "tosca Template description"
      - name: "id"
        in: "path"
        description: "ID of topolog template to return"
        required: true
        type: "string"       
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
        405:
          description: "Invalid input"          
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
    delete:
      summary: "Deletes a tosca topolog template"
      description: ""
      operationId: "deleteToscaTemplateByID"
      parameters:
      - name: "id"
        in: "path"
        description: "ID of topolog template to return"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        400:
          description: "Invalid ID supplied"
        404:
          description: "ToscaTemplate not found"
      security:
      - drip_auth:
        - "write:ToscaTemplate"
        - "read:ToscaTemplate"
...
  /user:
    post:
      summary: "Create user"
      description: "This can only be done by admin."
      operationId: "createUser"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
        409:
          description: "already exists"
      security:
      - drip_auth:
        - "admin:User"
  /user/login:
    get:
      tags:
      - "user"
      summary: "Logs user into the system"
      description: ""
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "password"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "string"
          headers:
            X-Rate-Limit:
              type: "integer"
              format: "int32"
              description: "calls per hour allowed by the user"
            X-Expires-After:
              type: "string"
              format: "date-time"
              description: "date in UTC when token expires"
        400:
          description: "Invalid username/password supplied"
...
  /user/{username}:
    get:
      summary: "Get user by user name"
      description: ""
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
      security:
      - drip_auth:
        - "admin:User"          
    put:
      tags:
      - "user"
      summary: "Updated user"
      description: "This can only be done by the logged in user."
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "name that need to be updated"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        400:
          description: "Invalid user supplied"
        404:
          description: "User not found"
      security:
      - drip_auth:
        - "admin:User"    
    delete:
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      produces:
      - "application/json"
      parameters:
      - name: "username"
        in: "path"
        description: "The name that needs to be deleted"
        required: true
        type: "string"
      responses:
        400:
          description: "Invalid username supplied"
        404:
          description: "User not found"
      security:
      - drip_auth:
        - "admin:User"      
securityDefinitions:
  drip_auth:
    type: "oauth2"
    authorizationUrl: "http://localhost/oauth/dialog"
    flow: "implicit"
    scopes:
      write:ToscaTemplate: "modify topolog template in your account"
      read:ToscaTemplate: "read your topolog template"
      admin:User: "Grants access to admin operations"
      write:Credentials: "modify cloud credentials in your account"
      read:Credentials: "read your cloud credentials"
definitions:
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
...